Blacklist a certain name from generating.
It's still in the training data to make sure it works, but really... it's a markov chain. If there are any words in the training data that contain letters in the necessary order, "Ladouceur" can be generated out of pure chance. This code change blacklists that so certain people in my real life don't get angry at me when their full name generates as an NPC.